\name{apmsTools-package}
\alias{apmsTools-package}
\alias{apmsTools}
\docType{package}
\title{
Tools for analysis of Affinity Purification Mass Spectrometry data
}
\description{
Several functions for handling/analyzing Spectral Count data from Affinity Purification Mass Spectrometry experiments
}
\details{
\tabular{ll}{
Package: \tab apmsTools\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2014-04-01\cr
License: \tab GPL-3\cr
Packaged: \tab 2014-04-04 09:26:40 UTC; erikahrne\cr
Built: \tab R 3.0.1; ; 2014-04-04 09:26:45 UTC; unix\cr
}
}
\author{
Erik Ahrne

Maintainer: Erik Ahrne <erik.ahrne@unibas.ch>

}
\references{
Not yet available
}
\keyword{ package }
\seealso{
Not yet available
}
\examples{
}
