\name{writeCRAPomeFile}
\alias{writeCRAPomeFile}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
	 Write CRAPome input file  
	 
}
\description{
Write CRAPome input file  

Example:
	 
PROTID	GENEID	PROTLEN	C1_SPC	C2_SPC	HDAC5_1_SPC	HDAC5_2_SPC

PROTID	GENEID	PROTLEN	C	C	HDAC5	HDAC5
	 
Q9Y6M1	IGF2BP2	599	9	2	2	0
	 
Q9Y697	NFS1	457	0	0	14	8

Q9Y618	NCOR2	2525	0	0	19	16

}
\usage{
writeCRAPomeFile(apmsExp, file, verbose = F)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{apmsExp}{
An object of class apmsExp.
}
  \item{file}{
Full path to output file.
}
  \item{verbose}{
verbose
}
}
\details{
Not yet available
}
\references{
Not yet available
}
\author{
Erik Ahrne

erik.ahrne@unibas.ch
}
\note{
Not yet available
}
\seealso{
Not yet available
}
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
