\name{writeSaintBaitFile}
\alias{writeSaintBaitFile}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Write bait SAINT input file 
}
\description{
Write bait SAINT input file 

Example:

	PPP2R1A	PPP2R1A	T
	
	PPP2CB	PPP2CB	T
	
	PPME1	PPME1	T
	
	TP53	TP53	T
	
	GFP1	GFP1	C
	
	GFP2	GFP2	C
	
	GFP3	GFP3	C

	GFP4	GFP4	C

}
\usage{
writeSaintBaitFile(apmsExp, file, verbose = F)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{apmsExp}{
An object of class apmsExp.
}
  \item{file}{
Full path to output file.
}
  \item{verbose}{
verbose
}
}
\details{
Not yet available
}
\references{
Not yet available
}
\author{
Erik Ahrne

erik.ahrne@unibas.ch
}
\note{
Not yet available
}

\seealso{
Not yet available
}
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line