\name{writeSaintInteractionFile}
\alias{writeSaintInteractionFile}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Write interaction SAINT input file.
}
\description{
Write interaction SAINT input file.

Example:

	PPP2R1A	PPP2R1A	sp|P30153|2AAA_HUMAN	64
	
	PPP2R1A	PPP2R1A	sp|Q9BUJ2|HNRL1_HUMAN	62
	
	PPP2R1A	PPP2R1A	sp|P08107|HSP71_HUMAN	29
}
\usage{
writeSaintInteractionFile(apmsExp, file, verbose = F)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{apmsExp}{
An object of class apmsExp.
}
  \item{file}{
Full path to output file.
}
  \item{verbose}{
verbose
}
}
\details{
Not yet available
}
\references{
Not yet available
}
\author{
Erik Ahrne

erik.ahrne@unibas.ch
}
\note{
Not yet available
}

\seealso{
Not yet available
}
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
